pyproject.toml
[tool.poetry]
name = "Sql_Python"
version = "0.1.0"
description = ""
authors = ["largolo <luciano.argolo@mercadolibre.com>"]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.26.68"
pandas = "^1.5.3"
pyspark = "^3.3.1"
requests = "^2.28.2"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
